upload all class

+++ src/main/java/com/example/test_2/entity/ClassEntity.java
@@ -16,7 +16,7 @@ import java.util.List;
 public class ClassEntity {
     @Id
     @Column(name = "class_id")
-    private int classId;
+    private Long classId;
 
     @Column(name = "class_name")
     private String className;
