+++ src/main/java/com/example/test_2/repository/AcademicResultRepository.java
@@ -3,5 +3,5 @@ package com.example.test_2.repository;
 import com.example.test_2.entity.AcademicResultEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AcademicResultRepository extends JpaRepository <AcademicResultEntity, Integer> {
+public interface AcademicResultRepository extends JpaRepository <AcademicResultEntity, Long> {
 }
