+++ src/main/java/com/example/test_2/repository/ClassRepository.java
@@ -3,5 +3,5 @@ package com.example.test_2.repository;
 import com.example.test_2.entity.ClassEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClassRepository extends JpaRepository<ClassEntity, Integer>{
+public interface ClassRepository extends JpaRepository<ClassEntity, Long>{
 }
