+++ src/main/java/com/example/test_2/repository/MedicalRepository.java
@@ -3,5 +3,5 @@ package com.example.test_2.repository;
 import com.example.test_2.entity.MedicalEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MedicalRepository extends JpaRepository<MedicalEntity, Integer> {
+public interface MedicalRepository extends JpaRepository<MedicalEntity, Long> {
 }
