+++ src/main/java/com/example/test_2/repository/StudentRepository.java
@@ -3,5 +3,5 @@ package com.example.test_2.repository;
 import com.example.test_2.entity.StudentEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StudentRepository extends JpaRepository<StudentEntity, Integer> {
+public interface StudentRepository extends JpaRepository<StudentEntity, Long> {
 }
