+++ src/main/java/com/example/test_2/repository/TuitionRepository.java
@@ -3,5 +3,5 @@ package com.example.test_2.repository;
 import com.example.test_2.entity.TuitionEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TuitionRepository extends JpaRepository<TuitionEntity, Integer> {
+public interface TuitionRepository extends JpaRepository<TuitionEntity, Long> {
 }
